+++ app/controllers/v2/todos_controller.rb
@@ -1,0 +1,5 @@
+class V2::TodosController < ApplicationController
+  def index
+    json_response({ message: 'Hello there'})
+  end
+end
